vc: start using qemu:///embed instead of qemu:///system
There are some problems associated with using the standard URI:
one of them is that it implies having libvirtd running at all
times, which is not very well liked in the container world, but
a much more critical one is that some of the devices we need to
access, chiefly the TAP device associated with an Endpoint, are
created inside namespaces that libvirtd doesn't have access to.
With qemu:///embed, libvirt is not running in a separate process,
so that's no longer a problem.