vc: recover libvirtRoot for subsequent uses
kata-runtime is a short-lived process, and the way it can retain
state between invocations is by having a directory on the host
filesystem where all necessary information is stored: the name of
this directory contains the sandbox ID, which is also passed to
all methods.

For libvirt information, and specifically libvirtRoot, we need to
do the same, however we only have the sandbox ID to work with. The
solution is to create a symlink from the regular kata-runtime
state directory to the corresponding libvirtRoot, and use that to
recover state for all invocations after the first one.